+++ packages/cohere/src/cohere-chat-language-model.ts
@@ -5,6 +5,7 @@ import {
   UnsupportedFunctionalityError,
 } from '@ai-sdk/provider';
 import {
+  FetchFunction,
   ParseResult,
   combineHeaders,
   createJsonResponseHandler,
@@ -26,7 +27,7 @@ type CohereChatConfig = {
   baseURL: string;
   headers: () => Record<string, string | undefined>;
   generateId: () => string;
-  fetch?: typeof fetch;
+  fetch?: FetchFunction;
 };
 
 export class CohereChatLanguageModel implements LanguageModelV1 {