+++ packages/provider-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ai-sdk/provider-utils
 
+## 1.0.12
+
+### Patch Changes
+
+- dd712ac: fix: use FetchFunction type to prevent self-reference
+
 ## 1.0.11
 
 ### Patch Changes