+++ packages/solid/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/solid
 
+## 0.0.35
+
+### Patch Changes
+
+- Updated dependencies [dd712ac]
+  - @ai-sdk/provider-utils@1.0.12
+  - @ai-sdk/ui-utils@0.0.32
+
 ## 0.0.34
 
 ### Patch Changes